Move Template project to the separate folder
