+++ src/Shark.Fido2.sln
@@ -53,12 +53,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.PostgreSql", "S
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.InMemory.Tests", "..\tests\Shark.Fido2.InMemory.Tests\Shark.Fido2.InMemory.Tests.csproj", "{D0880235-D8C8-04BD-41F9-3B8EE4C6A911}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.VisualStudio.Template", "Shark.Fido2.VisualStudio.Template\Shark.Fido2.VisualStudio.Template.csproj", "{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Templates", "Templates", "{9932CECD-B487-4BB5-89C1-62B33F8D0175}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.Sample.VisualStudio.Template", "Shark.Fido2.VisualStudio.Template\content\Shark.Fido2.Sample.VisualStudio.Template.csproj", "{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.ConvenienceMetadata.Core", "Shark.Fido2.ConvenienceMetadata.Core\Shark.Fido2.ConvenienceMetadata.Core.csproj", "{27CA81BA-F81D-44A9-9064-EF5F2426291F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.ConvenienceMetadata.Core.Tests", "..\tests\Shark.Fido2.ConvenienceMetadata.Core.Tests\Shark.Fido2.ConvenienceMetadata.Core.Tests.csproj", "{E8084625-0099-4CA9-892C-A8FAE55B3455}"
@@ -339,30 +333,6 @@ Global
 		{D0880235-D8C8-04BD-41F9-3B8EE4C6A911}.Release|x64.Build.0 = Release|Any CPU
 		{D0880235-D8C8-04BD-41F9-3B8EE4C6A911}.Release|x86.ActiveCfg = Release|Any CPU
 		{D0880235-D8C8-04BD-41F9-3B8EE4C6A911}.Release|x86.Build.0 = Release|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Debug|x64.Build.0 = Debug|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Debug|x86.Build.0 = Debug|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Release|x64.ActiveCfg = Release|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Release|x64.Build.0 = Release|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Release|x86.ActiveCfg = Release|Any CPU
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Release|x86.Build.0 = Release|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Debug|x64.Build.0 = Debug|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Debug|x86.Build.0 = Debug|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Release|x64.ActiveCfg = Release|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Release|x64.Build.0 = Release|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Release|x86.ActiveCfg = Release|Any CPU
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Release|x86.Build.0 = Release|Any CPU
 		{27CA81BA-F81D-44A9-9064-EF5F2426291F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{27CA81BA-F81D-44A9-9064-EF5F2426291F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{27CA81BA-F81D-44A9-9064-EF5F2426291F}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -421,8 +391,6 @@ Global
 		{B9AF26AE-1570-4490-B396-12AF051B506A} = {0DBAB265-1DF6-444B-BC5F-7A47857979BB}
 		{9BED5EF2-38EC-AADA-5FBA-4F9D33D1EC28} = {06FE717D-4AC9-4959-A179-260EE0DBCB27}
 		{D0880235-D8C8-04BD-41F9-3B8EE4C6A911} = {0DBAB265-1DF6-444B-BC5F-7A47857979BB}
-		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3} = {9932CECD-B487-4BB5-89C1-62B33F8D0175}
-		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723} = {9932CECD-B487-4BB5-89C1-62B33F8D0175}
 		{E8084625-0099-4CA9-892C-A8FAE55B3455} = {0DBAB265-1DF6-444B-BC5F-7A47857979BB}
 		{78F76A91-8746-484C-B924-4B6E6FF60F9D} = {0DBAB265-1DF6-444B-BC5F-7A47857979BB}
 	EndGlobalSection
