+++ src/Shark.Fido2.slnx
@@ -10,10 +10,6 @@
     <Project Path="Shark.Fido2.PostgreSql/Shark.Fido2.PostgreSql.csproj" />
     <Project Path="Shark.Fido2.SqlServer/Shark.Fido2.SqlServer.csproj" />
   </Folder>
-  <Folder Name="/Templates/">
-    <Project Path="Shark.Fido2.VisualStudio.Template/content/Shark.Fido2.Sample.VisualStudio.Template.csproj" />
-    <Project Path="Shark.Fido2.VisualStudio.Template/Shark.Fido2.VisualStudio.Template.csproj" />
-  </Folder>
   <Folder Name="/Tests/">
     <Project Path="../tests/Shark.Fido2.Common.Tests/Shark.Fido2.Common.Tests.csproj" />
     <Project Path="../tests/Shark.Fido2.ConvenienceMetadata.Core.Tests/Shark.Fido2.ConvenienceMetadata.Core.Tests.csproj" />
