+++ templates/projects/Shark.Fido2.VisualStudio.Template.slnx
@@ -1,0 +1,6 @@
+<Solution>
+  <Folder Name="/Template/">
+    <Project Path="Shark.Fido2.VisualStudio.Template/content/Shark.Fido2.Sample.VisualStudio.Template.csproj" />
+    <Project Path="Shark.Fido2.VisualStudio.Template/Shark.Fido2.VisualStudio.Template.csproj" />
+  </Folder>
+</Solution>
