+++ templates/projects/Shark.Fido2.VisualStudio.Template/Shark.Fido2.VisualStudio.Template.csproj
