+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/.template.config/template.json
