+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/Pages/CredentialDetails.cshtml
@@ -4,7 +4,7 @@
 <div class="main">
     <h1>Public Key Credentials</h1>
 
-    @if (Model.CredentialId == null || Model.CredentialId.Length == 0)
+    @if (Model.CredentialId.Length == 0)
     {
         <div class="section">
             <div class="alert alert-warning">
