+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/Pages/Index.cshtml.cs
