+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/Pages/Shared/_Layout.cshtml
