+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/Pages/_ViewImports.cshtml
