+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/Program.cs
