+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/Properties/launchSettings.json
