+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/Services/ICredentialService.cs
