+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/Shark.Fido2.Sample.VisualStudio.Template.csproj
