+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/appsettings.json
