+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/wwwroot/css/site.css
