+++ templates/projects/Shark.Fido2.VisualStudio.Template/content/wwwroot/js/attestation.js
